separate fluent ui tasks
