+++ package.json
@@ -45,11 +45,11 @@
     "postinstall": "node ./scripts/postinstall.js",
     "prettier": "node scripts/prettier.js",
     "rebuild": "node ./scripts/invalidate-just-cache.js && yarn build --reset-cache",
-    "release:fluentui:minor": "yarn workspace @fluentui/scripts just fluentui:publish:minor",
-    "release:fluentui:patch": "yarn workspace @fluentui/scripts just fluentui:publish:patch",
-    "release:fluentui:canary": "yarn workspace @fluentui/scripts just fluentui:publish:canary",
-    "release:fluentui:post-validation": "yarn workspace @fluentui/scripts just fluentui:publish:validation",
-    "release:fluentui:pack-nightly": "yarn workspace @fluentui/scripts just fluentui:pack",
+    "release:fluentui:minor": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-minor",
+    "release:fluentui:patch": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-patch",
+    "release:fluentui:canary": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-canary",
+    "release:fluentui:post-validation": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish post-publish",
+    "release:fluentui:pack-nightly": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish pack-nightly",
     "rename-package": "node -r ./scripts/ts-node-register ./scripts/rename-package.ts",
     "runto:lerna": "node ./scripts/monorepo/runTo.js",
     "run:published": "node ./scripts/monorepo/runPublished.js",
